+++ layout/src/main/java/com/hortonworks/iotas/layout/design/rule/condition/ExpressionVisitor.java
@@ -10,4 +10,5 @@ interface ExpressionVisitor {
     void visit(MapFieldExpression mapFieldExpression);
     void visit(Literal literal);
     void visit(FunctionExpression functionExpression);
+    void visit(AggregateFunctionExpression aggregateFunctionExpression);
 }