+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 ## 0.1.3
+ * IOT-269: Normalization processor runtime and UI integration
+ * IOT-239: TP-3 release bug fixes
+ * IOT-247: Add integration test for tagging
+ * IOT-8: Sliding Time Windows
+ * IOT-252: Add integration test for StreamInfo storage with JDBC provider
+ * IOT-261: topology status API can't determine status of topology on Storm
+ * IOT-262: Fix tagging api throwing java.lang.StackOverflowError: null
+ * IOT-259: Do not throw an exception if schema is null
+ * IOT-256: Hierarchical tagging UI integration
+ * IOT-255: UI integration to improve parser upload experience
  * IOT-265: Topologies with notification sinks throw error while deploying to storm 
  * IOT-248: Integration tests for Notification Service
  * IOT-189: Support adding custom artifacts in storm.jar